add functional for edit news and posts page in amdin panel

+++ application/models/Admin.php
@@ -5,13 +5,15 @@ namespace Application\Models;
 use Application\Core\Model;
 use \Imagick;
 
-class Admin extends Model {
+class Admin extends Model
+{
 
     public $error = [];
 
-    public function loginValidate($post) {
+    public function loginValidate($post)
+    {
         $config = require "application/config/admin.php";
-        if($config['login'] != $post['login'] or $config['password'] != $post['password']) {
+        if ($config['login'] != $post['login'] or $config['password'] != $post['password']) {
             $this->error[] = "Логин или пароль указан неверно";
             return false;
         }
@@ -19,7 +21,8 @@ class Admin extends Model {
         return true;
     }
 
-    public function postValidate($type, $post) {
+    public function postValidate($type, $post)
+    {
         $titleLength = mb_strlen($post['title']);
         $descriptionLength = mb_strlen($post['description']);
         $contentLength = mb_strlen($post['content']);
@@ -35,33 +38,14 @@ class Admin extends Model {
             $this->error['content'] = "Контент поста должен содержать не менее 30ти символов";
         }
 
-        $img = $_FILES['img']['tmp_name'];
-        if ($type == 'edit' and empty($img)) {
-            if(empty($img)) {
-                $this->error['img'] = "Изображение не выбрано";
-            }
-        }
-
-        if(empty($this->error)) {
+        if (empty($this->error)) {
             return true;
         }
         return false;
     }
 
-//    public function addPost($id, $post) {
-//        $params = [
-//            "title" => $post['title'],
-//            "description" => $post['description'],
-//            "content" => $post['content'],
-//            "status" => $post['status'],
-//            "publication_date" => strtotime(date("Y-m-d H:i:s")),
-//        ];
-//
-//        $this->db->query("INSERT INTO posts (title, description, content, status, publication_date) VALUES (:title, :description, :content, :status, :publication_date)", $params);
-//        return $this->db->lastInsertId();
-//    }
-
-    public function addPost($post) {
+    public function addPost($post)
+    {
         $params = [
             "title" => $post['title'],
             "description" => $post['description'],
@@ -74,41 +58,54 @@ class Admin extends Model {
         return $this->db->lastInsertId();
     }
 
-    public function postUploadImage($file, $id) {
-        move_uploaded_file($file, 'public/uploads/' . $id . '.jpg');
-    }
-
-    public function isPostExists($id) {
-        $params = [
-            'id' => $id,
-        ];
-        return $this->db->column('Select id from posts where id = :id', $params);
-    }
-
-    public function deletePost($id) {
-        $params = [
-            'id' => $id
-        ];
-        $this->db->query("DELETE FROM posts WHERE id = :id", $params);
-        unlink('public/uploads/' . $id . '.jpg');
-    }
-
-    public function getPost($id) {
-        $params = [
-            'id' => $id
-        ];
-        return $this->db->row("SELECT * FROM posts WHERE id = :id", $params);
-    }
-
-    public function updatePost($id, $post) {
+    public function updatePost($id, $post)
+    {
         $params = [
             'id' => $id,
             "title" => $post['title'],
             "description" => $post['description'],
             "content" => $post['content'],
             "status" => $post['status'],
-            "publication_date" => strtotime(date("Y-m-d H:i:s")),
+            "last_edit_date" => strtotime(date("Y-m-d H:i:s")),
         ];
-        return $this->db->query("UPDATE posts SET title = :title, description = :description, content = :content, status = :status, publication_date = :publication_date) where id = :id", $params);
+        return $this->db->query("UPDATE posts SET title = :title, description = :description, content = :content, status = :status, last_edit_date = :last_edit_date WHERE id = :id", $params);
     }
+
+    public function postUploadImage($file, $id)
+    {
+        move_uploaded_file($file, 'public/uploads/' . $id . '.jpg');
+    }
+
+    public function isPostExists($id)
+    {
+        $params = [
+            'id' => $id,
+        ];
+        return $this->db->column('Select id from posts where id = :id', $params);
+    }
+
+    public function deletePost($id)
+    {
+        $params = [
+            'id' => $id
+        ];
+        $this->db->query("DELETE FROM posts WHERE id = :id", $params);
+        if (file_exists('public/uploads/' . $id . '.jpg')) {
+            unlink('public/uploads/' . $id . '.jpg');
+        }
+    }
+
+    public function getPost($id)
+    {
+        $params = [
+            'id' => $id
+        ];
+        return $this->db->row("SELECT * FROM posts WHERE id = :id", $params);
+    }
+
+    public function getAllPosts()
+    {
+        return $this->db->row("SELECT * FROM posts");
+    }
+
 }
